Enables remaining tests
